minesweeper_env_rl/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minesweeper-env-rl"
version = "0.1.0"
description = "JAX Minesweeper environment for reinforcement learning agents."
requires-python = ">=3.10"
keywords = ["reinforcement-learning", "jax", "minesweeper"]
dependencies = [
    "altair==5.5.0",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "cachetools==6.2.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "fonttools==4.60.1",
    "gitdb==4.0.12",
    "GitPython==3.1.45",
    "idna==3.10",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "jax==0.7.2",
    "jaxlib==0.7.2",
    "Jinja2==3.1.6",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "kiwisolver==1.4.9",
    "MarkupSafe==3.0.3",
    "matplotlib==3.10.6",
    "ml_dtypes==0.5.3",
    "narwhals==2.7.0",
    "numpy==2.3.3",
    "opt_einsum==3.4.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==11.3.0",
    "protobuf==6.32.1",
    "pyarrow==21.0.0",
    "pydeck==0.9.1",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rpds-py==0.27.1",
    "scipy==1.16.2",
    "six==1.17.0",
    "smmap==5.0.2",
    "streamlit==1.50.0",
    "streamlit-image-coordinates==0.4.0",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5.2",
    "typing_extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "watchdog==6.0.0",
]

[tool.setuptools.packages.find]
include = ["minesweeper_env_rl*"]
